deleting of unnecessary folders

+++ run.py
@@ -1,5 +1,6 @@
 from flask import Flask, render_template, request
 from before_dashboard_code import compute_price_increase
+
 app = Flask(__name__)
 
 
@@ -11,12 +12,13 @@ def home():
 @app.route("/result", methods=['POST'])
 def result():
     if request.method == 'POST':
-        distance = int(request.form['distance'])
-        age = int(request.form['age'])
-        population = int(request.form['population'])
+        distance = float(request.form['distance'])
+        age = float(request.form['age'])
+        population = float(request.form['population'])
         energy_rating = int(request.form['energy_rating'])
         result = compute_price_increase(
             distance, age, population, energy_rating)
+
         return render_template("result.html", result=result)
